Revert "Rename file_manager_s3 module to file_manager"
This reverts commit 4daa04c545d47fc00d6c42a6ab0a73775944bacd.
